bugfix: on skier crash push left cause game crash
the problem was that in 'setupKeyHandler' function, if pressing 'left',
and the current skierDirection is not 1(left), then it was decrementing
skierDirection. If a skier collision had occurred, skierDirection would
be 0, which would result in skierDirection === -1, which is an invalid
value, causing a non-existent skierImage to be loaded in
'checkIfSkierHitObstacle()'. The crash would happen during generation of
skierRect right limit - skierImage.width would be undefined.